requirements.txt
Flask==1.1.2
Flask-Cors==3.0.8
gunicorn==20.0.4
python-chess==0.31.3
tensorflow==2.3.0
tensorflow-addons==0.11.1
numpy==1.18.5
urllib3==1.25.10